requirements.txt


# === Core Dependencies ===
python-dotenv>=1.0.0
playwright>=1.49.0

# === Optional Utilities ===
colorama>=0.4.6
imapclient>=3.0.0
email-validator>=2.1.0.post1
